ProductScreenAdaptive (phone view) 99% done
