+++ package.json
@@ -7,7 +7,6 @@
     "@testing-library/react": "^11.2.6",
     "@testing-library/user-event": "^12.8.3",
     "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
     "dotenv": "^8.2.0",
     "formik": "^2.2.6",
     "jwt-decode": "^3.1.2",
