+++ src/App.js
@@ -1,14 +1,12 @@
 import React from 'react';
 import AppRoutes from './routes/AppRoutes';
 import Header from './components/Header/Header';
-import Footer from './components/Footer/Footer';
 
 function App() {
   return (
     <div>
       <Header />
       <AppRoutes />
-      <Footer />
     </div>
   );
 }
