+++ src/components/SlickSlider/SlickSlider.scss
@@ -1,15 +1,7 @@
-.add-to-cart {
-  background-color: aqua;
-  font-size: 1rem;
-  color: azure;
-  padding: 4px 8px;
-  border-radius: 3px;
-  border: none;
-  margin-top: 10px;
-}
+
 .img {
-  max-height: 190px;
-  width: 150px
+  max-height: 130px;
+  width: 75px;
 }
 .name {
   font-size: 12px;
@@ -28,16 +20,19 @@ body {
   position: absolute;
   top: 35%;
   z-index: 1;
+  display: none;
 
 }
 
 .slick-prev:before {
   font-size: 35px;
   color: #696969;
+  
 
 }
 .slick-next:before {
   font-size: 35px;
   color: 	#696969;
+
 
 }
