--- src/pages/ProductScreen/productscreen-components/WarrantyInfo.module.scss
@@ -1,3 +1,0 @@
-.warranty {
-    margin-top: 20px;
-}
