+++ src/pages/ProductScreenAdaptive/productscreen-components/DeliveryInfo.module.scss
@@ -1,10 +1,11 @@
 .li {
     display: list-item;
     list-style-type: "—";
+    font-size: 0.8rem;
     padding-inline-start: 1ch;
     margin-bottom: 10px;
 }
 .ul {
     margin-left: 20px;
-    margin-top: 20px;
+    margin-top: 10px;
 }
