+++ src/pages/ProductScreenAdaptive/productscreen-components/PaymentInfo.module.scss
@@ -1,0 +1,7 @@
+.li {
+    display: list-item;
+    list-style-type: "—";
+    font-size: 0.8rem;
+    padding-inline-start: 1ch;
+    margin-bottom: 10px;
+}
