+++ src/pages/ProductScreenAdaptive/productscreen-components/WarrantyInfo.module.scss
@@ -1,0 +1,5 @@
+.warranty {
+    margin-top: 10px;
+    font-size: 0.8rem;
+
+}
